+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added OAuth 2.0 error middleware to `/authorize` endpoint.
+
 ### Changed
 - `AccessTokenService#issue()` is called with `msg` argument containing `scope`
 property, rather than `grant.scope`, when called from token response type.